Move AbstractXmlConverterTest to src to be used in other repositories
