+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/BusBreakerAnonymiserBugTest.java
@@ -11,6 +11,7 @@ import com.powsybl.commons.config.PlatformConfig;
 import com.powsybl.commons.datasource.MemDataSource;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.VoltageLevel;
+import com.powsybl.iidm.xml.test.AbstractXmlConverterTest;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
