+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/ReactiveLimitsXmlTest.java
@@ -7,6 +7,7 @@
 package com.powsybl.iidm.xml;
 
 import com.powsybl.iidm.network.test.ReactiveLimitsTestNetworkFactory;
+import com.powsybl.iidm.xml.test.AbstractXmlConverterTest;
 import org.junit.Test;
 
 import java.io.IOException;
