+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/SkipExtensionTest.java
@@ -8,6 +8,7 @@
 package com.powsybl.iidm.xml;
 
 import com.powsybl.iidm.network.Network;
+import com.powsybl.iidm.xml.test.AbstractXmlConverterTest;
 import org.junit.Test;
 
 import java.io.IOException;
