+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/extensions/HvdcAngleDroopActivePowerControlXmlTest.java
@@ -19,7 +19,7 @@ import org.junit.Test;
 
 import java.io.IOException;
 
-import static com.powsybl.iidm.xml.AbstractXmlConverterTest.getVersionedNetworkPath;
+import static com.powsybl.iidm.xml.test.AbstractXmlConverterTest.getVersionedNetworkPath;
 import static com.powsybl.iidm.xml.IidmXmlConstants.CURRENT_IIDM_XML_VERSION;
 import static org.junit.Assert.assertEquals;
