+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/extensions/SubstationAndLinePositionXmlTest.java
@@ -17,7 +17,7 @@ import org.junit.Test;
 import java.io.IOException;
 import java.util.List;
 
-import static com.powsybl.iidm.xml.AbstractXmlConverterTest.getVersionedNetworkPath;
+import static com.powsybl.iidm.xml.test.AbstractXmlConverterTest.getVersionedNetworkPath;
 import static com.powsybl.iidm.xml.IidmXmlConstants.CURRENT_IIDM_XML_VERSION;
 import static org.junit.Assert.assertNotNull;
